added tests
added test for checkDuration (looks at metadata diff)

added test for checkFrameRate
