+++ tests/test_video_tools.py
@@ -307,14 +307,6 @@ class TestVideoTools(TestSupport):
         rate = locator.get_frame_rate(audio=True)
         self.assertTrue(abs(rate - 44100) < 1)
 
-    def test_meta(self):
-        meta, frames = get_meta_from_video(self.locateFile('tests/videos/sample1.mov'), with_frames=True)
-        self.assertEqual(803, len(frames[0]))
-        self.assertEqual(2557, len(frames[1]))
-        meta, frames = get_meta_from_video(self.locateFile('tests/videos/sample1.mov'), show_streams=True)
-        self.assertEqual('yuv420p', meta[0]['pix_fmt'])
-        self.assertEqual('audio', meta[1]['codec_type'])
-
     def test_frame_count(self):
